Initial Commit
Addition of VoronoiLab, a Java program written as a solution to the
third homework assignment for COMPSCI 122 - Graphical User Interfaces in
Java with Dr. Drue Coles. VoronoiLab is intended as a visual interface
for creating simple voronoi diagrams by clicking within the component
area in the window, as well as providing other geometric visualizations.